PDJB-794: pr markups - formatting, improve test structure, readd missing string
