+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/summaryModels/propertyComplianceViewModels/PropertyComplianceViewModelFactoryTests.kt
@@ -25,7 +25,7 @@ class PropertyComplianceViewModelFactoryTests {
     }
 
     private val gasSafetyViewModelFactory = GasSafetyViewModelFactory(mock(), mockMessageSource)
-    private val electricalSafetyViewModelFactory = ElectricalSafetyViewModelServiceRedesign(mock(), mock())
+    private val electricalSafetyViewModelFactory = ElectricalSafetyViewModelFactory(mock(), mock())
     private val propertyComplianceViewModelFactory =
         PropertyComplianceViewModelFactory(
             gasSafetyViewModelFactory,
